% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_intensity.R
\name{adjust_intensity}
\alias{adjust_intensity}
\alias{adjust_intensity.formula}
\alias{adjust_intensity.default}
\title{Adjust spectral intensities}
\usage{
adjust_intensity(x, ...)

\method{adjust_intensity}{formula}(formula, data = NULL, ...)

\method{adjust_intensity}{default}(x, y, type = "none", make_relative = TRUE, ...)
}
\arguments{
\item{x}{wavenumber}

\item{formula}{formula}

\item{data}{data}

\item{y}{intensity}

\item{type}{type}

\item{make_relative}{make_relative}

\item{\ldots}{...}
}
\description{
description
}
\details{
If the uploaded spectrum is not in absorbance units, use this input to specify the units to convert from.Open Specy can adjust reflectance or transmittance spectra to Absorbance units using this drop down in the upload file tab. All of the preceding tabs assume that the data is in absorbance units so you should make the correction before continuing if needed. The transmittance adjustment uses the log10(1/T) calculation which does not correct for system and particle characteristics. The reflectance adjustment uses the Kubelka-Munk equation (1-R)2/(2*R). If none is selected, Open Specy assumes that the uploaded data is an absorbance spectrum.
}
\examples{
data("raman_hdpe")
adjust_intensity(intensity ~ wavenumber, data = raman_hdpe)

}
\seealso{
seealso
}
