% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_negative.R
\name{adjust_negative}
\alias{adjust_negative}
\title{Adjust negative values}
\usage{
adjust_negative(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one by
\code{as.vector(x, "numeric")}.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
This helper function converts a numeric values \code{x < 1} into values >= 1
by keeping absolute differences between values. If \code{na.rm} is
\code{TRUE}, missing values are removed before the computation proceeds.
}
\examples{
adjust_negative(-5:5)
adjust_negative(c(0.2, 0.7, 1))

}
\seealso{
\code{\link[base]{min}} for the calculation of minima
}
