% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_subtraction.R
\name{background_subtraction}
\alias{background_subtraction}
\title{Automated background subtraction for Raman spectroscopy}
\usage{
background_subtraction(formula, data = NULL, degree = 8, ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{degree}{the degree of the polynomial. Must be less than the number of
unique points when raw is false, as by default.}

\item{\ldots}{...}
}
\description{
This baseline correction routine iteratively finds the baseline of a spectrum
using a polynomial fitting.
}
\examples{
data("raman_pe")
background_subtraction(absorbance ~ wavenumber, data = raman_pe)

}
\seealso{
seealso
}
