% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{adjust_negative}
\alias{adjust_negative}
\alias{make_relative}
\title{Normalization of spectral data}
\usage{
adjust_negative(x, na.rm = FALSE)

make_relative(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one by
\code{as.vector(x, "numeric")}.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
\code{adjust_negative()} converts numeric values \code{x < 1} into values
>= 1 by keeping absolute differences between values.
\code{make_relative()} converts values \code{x} into relative values between
0 and 1.
If \code{na.rm} is \code{TRUE}, missing values are removed before the
computation proceeds.
}
\examples{
adjust_negative(c(-1000, -1, 0, 1, 10))
make_relative(-2:10)

}
\seealso{
\code{\link[base]{min}} for the calculation of minima
}
