% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_OpenSpecy.R
\name{head.OpenSpecy}
\alias{head.OpenSpecy}
\alias{print.OpenSpecy}
\title{Generic functions for \code{OpenSpecy} objects}
\usage{
\method{head}{OpenSpecy}(x, ...)

print.OpenSpecy(x, ...)
}
\arguments{
\item{x}{x.}

\item{\ldots}{args.}
}
\value{
return
}
\description{
Functions to
}
\details{
\code{as_OpenSpecy()}
}
\examples{
c()

}
\seealso{
seealso.
}
\author{
Zacharias Steinmetz, Win Cowger
}
