% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lib.R
\name{get_lib}
\alias{get_lib}
\title{Download spectral libraries}
\usage{
get_lib(which = c("ftir", "raman"), ...)
}
\arguments{
\item{which}{which}

\item{\ldots}{...}
}
\description{
Description
}
\examples{
\dontrun{
get_lib()
}

}
\seealso{
seealso
}
