% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_spectrum.R
\name{match_spectrum}
\alias{match_spectrum}
\alias{match_spectrum.formula}
\alias{match_spectrum.data.frame}
\alias{match_spectrum.default}
\title{Match spectra with reference library}
\usage{
match_spectrum(x, ...)

\method{match_spectrum}{formula}(formula, data = NULL, ...)

\method{match_spectrum}{data.frame}(x, ...)

\method{match_spectrum}{default}(
  x,
  y,
  library,
  which = NULL,
  type = "full",
  range = seq(0, 6000, 0.1),
  col_names = NULL,
  top_n = 100,
  ...
)
}
\arguments{
\item{x}{wavenumber}

\item{formula}{formula}

\item{data}{data}

\item{y}{intensity}

\item{library}{library}

\item{which}{which}

\item{type}{type}

\item{range}{range}

\item{col_names}{col_names}

\item{top_n}{top_n}

\item{\ldots}{...}
}
\description{
desc
}
\examples{
\dontrun{
data("raman_hdpe")

get_lib("raman")
spec_lib <- load_lib("raman")

match_spectrum(raman_hdpe, spec_lib, "raman", "full")
}

}
\seealso{
seealso
}
