% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_reports.R
\name{heatmap_OpenSpecy}
\alias{heatmap_OpenSpecy}
\alias{plot_OpenSpecy}
\alias{interactive_plot}
\title{Plots for Open Specy objects}
\usage{
heatmap_OpenSpecy(
  object,
  z = NULL,
  sn = NULL,
  cor = NULL,
  min_sn = NULL,
  min_cors = NULL,
  selected_spectrum = NULL
)

plot_OpenSpecy(x, selected_spectrum = 1, x2 = NULL, selected_spectrum2 = 1)

interactive_plot(x, selected_spectrum, x2, selected_spectrum2)
}
\arguments{
\item{object}{An OpenSpecy object containing metadata and spectral data.}

\item{z}{Optional numeric vector specifying the intensity values for the heatmap. If not provided, the function will use the intensity values from the OpenSpecy object.}

\item{sn}{Optional numeric value specifying the signal-to-noise ratio threshold. If provided along with \code{min_sn}, regions with SNR below the threshold will be excluded from the heatmap.}

\item{cor}{Optional numeric value specifying the correlation threshold. If provided along with \code{min_cors}, regions with correlation below the threshold will be excluded from the heatmap.}

\item{min_sn}{Optional numeric value specifying the minimum signal-to-noise ratio for inclusion in the heatmap. Regions with SNR below this threshold will be excluded.}

\item{min_cors}{Optional numeric value specifying the minimum correlation for inclusion in the heatmap. Regions with correlation below this threshold will be excluded.}

\item{selected_spectrum}{Optional index of the selected spectrum to plot.}

\item{x}{An OpenSpecy object containing metadata and spectral data for the first group.}

\item{x2}{An optional second OpenSpecy object containing metadata and spectral data for the second group.}

\item{selected_spectrum2}{Optional index of the selected spectrum from the second group to plot.}
}
\value{
A plotly heatmap object displaying the OpenSpecy data. A subplot containing the heatmap and spectra plot. A plotly object displaying the spectra from the OpenSpecy object(s).
}
\description{
These functions generate heatmaps, spectral plots, and interactive plots for OpenSpecy data.
}
\examples{
data1 <- read_any("raman_hdpe.json")
data2 <- read_zip("inst/extdata/CA_tiny_map.zip")
plot_OpenSpecy(data1)
plot_OpenSpecy(data1, selected_spectrum = 2)
plot_OpenSpecy(data1, x2 = data2, selected_spectrum2 = 3)
heatmap_OpenSpecy(data2, z = data2$metadata$y)
interactive_plot(data1, selected_spectrum = 2)
interactive_plot(data2, selected_spectrum = 2, x2 = data1, selected_spectrum2 = 1)
}
\author{
Win Cowger, Zacharias Steinmetz#'
}
