% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman_pe.R
\docType{data}
\name{raman_pe}
\alias{raman_pe}
\title{Sample Raman spectrum}
\format{
A data table containing 964 rows and 2 columns with information on:
\tabular{ll}{
wavenumber: \tab \cr
absorbance: \tab \cr
}
}
\description{
Polyethylene
}
\references{
ref
}
\author{
author
}
\keyword{data}
