% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{read_asp}
\alias{read_asp}
\alias{read_spa}
\alias{read_jdx}
\alias{read_spc}
\alias{read_0}
\title{Read spectral data}
\usage{
read_asp(file = ".", ...)

read_spa(file = ".", ...)

read_jdx(file = ".", ...)

read_spc(file = ".", ...)

read_0(file = ".", ...)
}
\arguments{
\item{file}{file.}

\item{\ldots}{...}
}
\description{
Read Raman or FTIR spectrum files as asp, jdx, spc, or spa. A csv file is
preferred. If a csv, the file must contain one column labeled Wavelength in
units of (1/cm) and another column labeled Absorbance in absorbance units.
If jdx, spc, spa, or 0 the file should be a single absorbance spectrum with
wavelength in (1/cm). These files will not always work perfectly because they
are tricky to read so double check them in another software.
}
\examples{
c()

}
\seealso{
seealso
}
