% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restric_range.R
\name{restrict_range}
\alias{restrict_range}
\alias{restrict_range.default}
\alias{restrict_range.OpenSpecy}
\title{Range Restriction for spectra}
\usage{
restrict_range(x, ...)

\method{restrict_range}{default}(object, ...)

\method{restrict_range}{OpenSpecy}(object, min_range = 0, max_range = 6000, make_rel = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one by
\code{as.vector(x, "numeric")}; \code{x} should contain the spectral
wavenumbers.}

\item{y}{a numeric vector containing the spectral intensities.}

\item{res}{spectral resolution supplied to \code{fun}.}

\item{fun}{the function to be applied to each element of \code{x}; defaults
to \code{\link[base]{round}()} to round to a specific resolution \code{res}.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
\code{adj_res()} abd \code{conform_res()} return a numeric vector with
resolution-conformed wavenumbers.
\code{adj_neg()} and \code{make_rel()} return numeric vectors
with the normalized intensity data.
}
\description{
\code{restrict_range()} restricts wavenumber ranges to user specified values.
Multiple ranges can be specified by inputting the series of max and min values in order.
}
\examples{
test_noise <- as_OpenSpecy(x = seq(400,4000, by = 10),
                           spectra = data.frame(intensity = rnorm(361)))
restrict_range(test_noise, min_range = 1000, max_range = 2000)
restrict_range(test_noise, min_range = c(1000, 2000),
               max_range = c(1500, 2500))

}
\seealso{
\code{\link[base]{min}()} and \code{\link[base]{round}()};
\code{\link{adj_intens}()} for log transformation functions
}
\author{
Win Cowger, Zacharias Steinmetz
}
