% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_spectra.R
\name{share_spectrum}
\alias{share_spectrum}
\alias{share_spectrum.default}
\alias{share_spectrum.data.frame}
\title{Share data with the Open Specy community}
\usage{
share_spectrum(data, ...)

\method{share_spectrum}{default}(data, ...)

\method{share_spectrum}{data.frame}(data, share = "system", ...)
}
\arguments{
\item{data}{a data frame containing the spectral data; columns should be
named \code{"wavenumber"} and \code{"intensity"}.}

\item{share}{\code{"system"} just keeps a local copy of the spectrum in the
Open Specy package directory at
\code{system.file("extdata", package = "OpenSpecy")} to be shared with us by
e-mail; if a correct API token exists, "dropbox" shares the spectrum with
the cloud.}

\item{\ldots}{further arguments passed to the submethods.}
}
\description{
This helper functions share spectra and metadata with the Open Specy
community.
}
\examples{
data("raman_hdpe")

share_spectrum(raman_hdpe)

}
\seealso{
\code{\link{read_text}()}
}
