% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_noise.R
\name{signal_noise}
\alias{signal_noise}
\title{Calculate signal and noise metrics for Open Specy objects}
\usage{
signal_noise(object, return = "signal_over_noise", na.rm = TRUE)
}
\arguments{
\item{object}{An Open Specy object.}

\item{return}{A character specifying the desired metric to calculate. Options include "signal" (mean intensity), "noise" (standard deviation of intensity), "signal_times_noise" (absolute value of signal times noise), "signal_over_noise" (absolute value of signal / noise), or "total_signal" (total signal = signal * number of data points).}

\item{na.rm}{Logical value indicating whether missing values should be removed when calculating signal and noise. Default is \code{TRUE}.}
}
\value{
A numeric vector containing the calculated metric for each spectrum in the Open Specy object.
}
\description{
This function calculates common signal and noise metrics for Open Specy objects.
}
\examples{
data("raman_hdpe")
signal_noise(os, return = "signal")
signal_noise(os, return = "noise")
signal_noise(os, return = "signal_times_noise")

}
