% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_spec.R
\name{split_spec}
\alias{split_spec}
\alias{split_spec.default}
\title{Split Open Specy objects}
\usage{
split_spec(x, ...)

\method{split_spec}{default}(x, ...)
}
\arguments{
\item{x}{a list of OpenSpecy objects or another supported spectral object.}

\item{...}{arguments passed to class methods.}
}
\value{
For ordinary inputs, a list of Open Specy objects each with one spectrum.
For \code{FileSpecs}, a named list of descriptor-only \code{FileSpecs} region views
that share the immutable source and cache.
}
\description{
Convert a list of Open Specy objects into one-spectrum objects, or split a
file-backed map into lightweight region views.
}
\details{
Function will accept a list of Open Specy objects of any length and will split
them to their individual components. For example a list of two objects,
an Open Specy with only one spectrum and an Open Specy with 50 spectra
will return a list of length 51 each with Open Specy objects that only have
one spectrum.
}
\examples{
data("test_lib")
data("raman_hdpe")
listed <- list(test_lib, raman_hdpe)
test <- split_spec(listed)
test2 <- split_spec(list(test_lib))

}
\seealso{
\code{\link{c_spec}()} for combining \code{OpenSpecy} objects.
\code{\link{collapse_spec}()} for summarizing \code{OpenSpecy} objects.
}
\author{
Zacharias Steinmetz, Win Cowger
}
