% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_app.R
\name{start_app}
\alias{start_app}
\title{Run Open Specy app}
\usage{
start_app(...)
}
\arguments{
\item{\dots}{arguments passed to \code{\link[shiny]{runApp}}.}
}
\description{
This function starts the graphical user interface of Open Specy.
}
\examples{
\dontrun{
start_app(launch.browser = TRUE)
}

}
