% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtr_bg.R
\name{subtr_bg}
\alias{subtr_bg}
\alias{subtr_bg.default}
\alias{subtr_bg.OpenSpecy}
\title{Automated background subtraction for spectral data}
\usage{
subtr_bg(x, ...)

\method{subtr_bg}{default}(x, ...)

\method{subtr_bg}{OpenSpecy}(
  x,
  degree = 8,
  wavenumber_fit,
  intensity_fit,
  raw = FALSE,
  make_rel = TRUE,
  type = "Polynomial",
  ...
)
}
\arguments{
\item{x}{a list object of class \code{OpenSpecy}.}

\item{degree}{the degree of the polynomial. Must be less than the number of
unique points when raw is \code{FALSE}. Typically a good fit can be
found with a 8th order polynomial.}

\item{wavenumber_fit}{manually specified wavenumbers for the baseline.}

\item{intensity_fit}{manually specified intensity values for the baseline.}

\item{raw}{if \code{TRUE}, use raw and not orthogonal polynomials.}

\item{make_rel}{logical; if \code{TRUE} spectra are automatically normalized
with \code{\link{make_rel}()}.}

\item{type}{one of "Polynomial" or "Manual" depending on whether you want spectra to be corrected with
a manual baseline or with polynomial baseline fitting.}

\item{\ldots}{further arguments passed to \code{\link[stats]{poly}()}.}
}
\value{
\code{subtr_bg()} returns a data frame containing two columns named
\code{"wavenumber"} and \code{"intensity"}.
}
\description{
This baseline correction routine iteratively finds the baseline of a spectrum
using a polynomial fitting or accepts a manual baseline.
}
\details{
This is a translation of Michael Stephen Chen's MATLAB code written for the
\code{imodpolyfit} routine from Zhao et al. 2007.
}
\examples{
data("raman_hdpe")

subtr_bg(raman_hdpe)

}
\references{
Chen MS (2020). Michaelstchen/ModPolyFit. \emph{MATLAB}.
Retrieved from \url{https://github.com/michaelstchen/modPolyFit}
(Original work published July 28, 2015)

Zhao J, Lui H, McLean DI, Zeng H (2007). “Automated Autofluorescence
Background Subtraction Algorithm for Biomedical Raman Spectroscopy.”
\emph{Applied Spectroscopy}, \strong{61}(11), 1225–1232.
\doi{10.1366/000370207782597003}.
}
\seealso{
\code{\link[stats]{poly}()};
\code{\link{smooth_intens}()}
}
\author{
Win Cowger, Zacharias Steinmetz
}
