% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_intens.R
\name{adj_intens}
\alias{adj_intens}
\alias{adj_intens.default}
\alias{adj_intens.OpenSpecy}
\title{Adjust spectral intensities to standard absorbance units.}
\usage{
adj_intens(x, ...)

\method{adj_intens}{default}(x, type = "none", make_rel = TRUE, log_exp = "none", ...)

\method{adj_intens}{OpenSpecy}(x, type = "none", make_rel = TRUE, log_exp = "none", ...)
}
\arguments{
\item{x}{a list object of class \code{OpenSpecy}.}

\item{type}{a character string specifying whether the input spectrum is
in absorbance units (\code{"none"}, default) or needs additional conversion
from \code{"reflectance"} or \code{"transmittance"} data.}

\item{make_rel}{logical; if \code{TRUE} spectra are automatically normalized
with \code{\link{make_rel}()}.}

\item{log_exp}{a character string specifying whether the input needs to be log
transformed \code{"log"}, exp transformed \code{"exp"}, or not (\code{"none"}, default).}

\item{\ldots}{further arguments passed to submethods; this is
to \code{\link{adj_neg}()} for \code{adj_intens()} and
to \code{\link{conform_res}()} for \code{conform_intens()}.}
}
\value{
\code{adj_intens()} returns a data frame containing two columns
named \code{"wavenumber"} and \code{"intensity"}.
}
\description{
Converts reflectance or transmittance intensity units to absorbance units and
adjust log or exp transformed units.
}
\details{
Many of the Open Specy functions will assume that the spectrum is in
absorbance units. For example, see \code{\link{subtr_baseline}()}.
To run those functions properly, you will need to first convert any spectra
from transmittance or reflectance to absorbance using this function.
The transmittance adjustment uses the \eqn{log(1 / T)}
calculation which does not correct for system and particle characteristics.
The reflectance adjustment uses the Kubelka-Munk equation
\eqn{(1 - R)^2 / 2R}. We assume that the reflectance intensity
is a percent from 1-100 and first correct the intensity by dividing by 100
so that it fits the form expected by the equation.
}
\examples{
data("raman_hdpe")

adj_intens(raman_hdpe)

}
\seealso{
\code{\link{subtr_baseline}()} for spectral background correction.
}
\author{
Win Cowger, Zacharias Steinmetz
}
