% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_intensity.R
\name{adjust_intensity}
\alias{adjust_intensity}
\alias{adjust_intensity.formula}
\alias{adjust_intensity.data.frame}
\alias{adjust_intensity.default}
\title{Adjust spectral intensities to absorbance units}
\usage{
adjust_intensity(x, ...)

\method{adjust_intensity}{formula}(formula, data = NULL, ...)

\method{adjust_intensity}{data.frame}(x, ...)

\method{adjust_intensity}{default}(x, y, type = "none", make_relative = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector containing the spectral wavenumbers; alternatively
a data frame containing spectral data as \code{"wavenumber"} and
\code{"intensity"} can be supplied.}

\item{formula}{an object of class '\code{\link[stats]{formula}}' of the form
\code{intensity ~ wavenumber}.}

\item{data}{a data frame containing the variables in \code{formula}.}

\item{y}{a numeric vector containing the spectral intensities.}

\item{type}{a character string specifying whether the input spectrum is
in absorbance units (\code{"none"}, default) or needs additional conversion
from \code{"reflectance"} or \code{"transmittance"} data.}

\item{make_relative}{logical; if \code{TRUE} spectra are automatically
normalized with \code{\link{make_relative}()}.}

\item{\ldots}{further arguments passed to the submethods.}
}
\description{
Converts reflectance or transmittance intensity units to absorbance units.
}
\details{
Many of the Open Specy functions will assume that the spectrum is in
absorbance units. For example, see \code{\link{match_spectrum}()} and
\code{\link{subtract_background}()}.
To run those functions properly, you will need to first convert any spectra
from transmittance or reflectance to absorbance using this function.
The transmittance adjustment uses the log10(1/T) calculation which does not
correct for system and particle characteristics.
\loadmathjax
The reflectance adjustment uses the Kubelka-Munk equation
\mjeqn{\frac{(1 - R)^2}{2R}}{(1 - R)2 / (2 * R)}.
}
\examples{
data("raman_hdpe")
adjust_intensity(raman_hdpe)

}
\seealso{
\code{\link{subtract_background}()} for spectral background correction;
\code{\link{match_spectrum}()} matches spectra with the Open Specy or other
reference libraries
}
