% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_under_band.R
\name{area_under_band}
\alias{area_under_band}
\alias{area_under_band.default}
\alias{area_under_band.OpenSpecy}
\title{Measure the area under band of spectra}
\usage{
area_under_band(x, ...)

\method{area_under_band}{default}(x, ...)

\method{area_under_band}{OpenSpecy}(x, min, max, na.rm = F, ...)
}
\arguments{
\item{x}{an \code{OpenSpecy} object.}

\item{min}{a numeric value of the smallest wavenumber to begin calculation.}

\item{max}{a numeric value of the largest wavenumber to end calculation.}

\item{na.rm}{a logical value for whether to ignore NA values.}

\item{\ldots}{additional arguments passed to vapply.}
}
\value{
Numeric vector of area under the band calculations for each spectrum in the Open Specy object.
}
\description{
Area under the band calculations are useful for quantifying spectral features.
Specialized fields in spectroscopy have different area under the band regions of interest
and their ratios that help with understanding differences in materials. Additional processing
is typically required prior to calculating these values for accuracy and reproducibility.
}
\examples{
data("raman_hdpe")
#Single area calculation
area_under_band(raman_hdpe, min = 1000,max = 2000)
#Ratio of two areas. 
area_under_band(raman_hdpe, min = 1000,max = 2000)/area_under_band(raman_hdpe, min = 500,max = 700)
}
\author{
Win Cowger
}
