% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_OpenSpecy.R
\name{as_OpenSpecy}
\alias{as_OpenSpecy}
\alias{as_OpenSpecy.data.frame}
\alias{as_OpenSpecy.default}
\alias{is_OpenSpecy}
\alias{OpenSpecy}
\alias{gen_grid}
\title{Coerce to Open Specy}
\usage{
as_OpenSpecy(x, ...)

\method{as_OpenSpecy}{data.frame}(x, colnames = list(wavenumber = NULL, spectra = NULL), ...)

\method{as_OpenSpecy}{default}(
  x,
  spectra,
  coords = "gen_grid",
  metadata = list(id = paste(digest(Sys.info()), digest(sessionInfo()), sep = "/"),
    file_name = NULL, user_name = NULL, contact_info = NULL, organization = NULL,
    citation = NULL, spectrum_type = NULL, spectrum_identity = NULL, material_form =
    NULL, material_phase = NULL, material_producer = NULL, material_purity = NULL,
    material_quality = NULL, material_color = NULL, material_other = NULL, cas_number =
    NULL, instrument_used = NULL, instrument_accessories = NULL, instrument_mode = NULL,
    spectral_resolution = NULL, 
     laser_light_used = NULL, number_of_accumulations =
    NULL, total_acquisition_time_s = NULL, data_processing_procedure = NULL,
    level_of_confidence_in_identification = NULL, other_info = NULL, license =
    "CC BY-NC"),
  ...
)

is_OpenSpecy(x)

OpenSpecy(x, ...)

gen_grid(x)
}
\arguments{
\item{x}{x.}

\item{colnames}{cols.}

\item{spectra}{spectra.}

\item{coords}{coords = "gen_grid".}

\item{metadata}{file = NULL.}

\item{\ldots}{args.}
}
\value{
return
}
\description{
Functions to check if an object is an Open Specy spectrum, or coerce it if
possible.
}
\details{
details
}
\examples{
c()

}
\seealso{
seealso.
}
\author{
Zacharias Steinmetz, Win Cowger
}
