% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_OpenSpecy.R
\name{as_OpenSpecy}
\alias{as_OpenSpecy}
\alias{as_OpenSpecy.data.frame}
\alias{as_OpenSpecy.default}
\alias{is_OpenSpecy}
\alias{OpenSpecy}
\alias{gen_grid}
\title{Coerce to Open Specy}
\usage{
as_OpenSpecy(x, ...)

\method{as_OpenSpecy}{data.frame}(x, colnames = list(wavenumber = NULL, spectra = NULL), ...)

\method{as_OpenSpecy}{default}(x, spectra, coords = "gen_grid", file = NULL)

is_OpenSpecy(x)

OpenSpecy(x, ...)

gen_grid(x)
}
\arguments{
\item{x}{x.}

\item{colnames}{cols.}

\item{spectra}{spectra.}

\item{coords}{coords = "gen_grid".}

\item{file}{file = NULL.}

\item{\ldots}{args.}
}
\value{
return
}
\description{
Functions to check if an object is an Open Specy spectrum, or coerce it if
possible.
}
\details{
details
}
\examples{
c()

}
\seealso{
seealso.
}
\author{
Zacharias Steinmetz, Win Cowger
}
