% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_subtraction.R
\name{background_subtraction}
\alias{background_subtraction}
\alias{background_subtraction.formula}
\alias{background_subtraction.default}
\title{Automated background subtraction for spectral data}
\usage{
background_subtraction(x, ...)

\method{background_subtraction}{formula}(formula, data = NULL, ...)

\method{background_subtraction}{default}(x, y, degree = 8, ...)
}
\arguments{
\item{x}{wavenumber}

\item{formula}{formula}

\item{data}{data}

\item{y}{absorbance}

\item{degree}{the degree of the polynomial. Must be less than the number of
unique points when raw is false, as by default.}

\item{\ldots}{...}
}
\description{
This baseline correction routine iteratively finds the baseline of a spectrum
using a polynomial fitting.
}
\examples{
data("raman_hdpe")
background_subtraction(absorbance ~ wavenumber, data = raman_hdpe)

}
\seealso{
seealso
}
