% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_particles.R
\name{collapse_spectra}
\alias{collapse_spectra}
\alias{characterize_particles}
\title{Characterize Particles}
\usage{
collapse_spectra(object)

characterize_particles(object, particles)
}
\arguments{
\item{object}{An OpenSpecy object}

\item{particles}{A logical vector or character vector describing which of the spectra are
of particles (TRUE) and which are not (FALSE). If a character vector is provided, it should
represent the different particle types present in the spectra.}
}
\value{
An Open Specy object appended metadata about the particles or collapsed for the particles.
}
\description{
Functions for analyzing particles in spectral map oriented OpenSpecy object.
}
\details{
`characterize_particles()` accepts an OpenSpecy object and a logical or character vector describing which pixels correspond to particles.
`collapse_spectra()` takes an OpenSpecy object with particle-specific metadata
(from `characterize_particles()`) and collapses the spectra to median intensities for each unique particle.
It also updates the metadata with centroid coordinates, while preserving the particle information on area and Feret max.
}
\examples{
#Logical example
map <- read_extdata("CA_tiny_map.zip") |> read_any()
map$metadata$particles <- map$metadata$x == 0
identified_map <- characterize_particles(map, map$metadata$particles)
test_collapsed <- collapse_spectra(identified_map)

#Character example
map <- read_extdata("CA_tiny_map.zip") |> read_any()
map$metadata$particles <- ifelse(map$metadata$x == 1, "particle", "not_particle")
identified_map <- characterize_particles(map, map$metadata$particles)
test_collapsed <- collapse_spectra(identified_map)

}
\author{
Win Cowger, Zacharias Steinmetz
}
