% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conform_spec.R
\name{conform_spec}
\alias{conform_spec}
\alias{conform_spec.default}
\alias{conform_spec.OpenSpecy}
\title{Conform spectra}
\usage{
conform_spec(x, new_wavenumbers, res = 5)

\method{conform_spec}{default}(x, new_wavenumbers, res = 5)

\method{conform_spec}{OpenSpecy}(x, new_wavenumbers, res = 5)
}
\arguments{
\item{x}{a list object of class \code{OpenSpecy}.}

\item{new_wavenumbers}{a vector of new wavenumber values, can be just supplied as a min and max value.}

\item{res}{spectral resolution adjusted to.}
}
\value{
\code{adj_intens()} returns a data frame containing two columns
named \code{"wavenumber"} and \code{"intensity"}.
}
\description{
Conform spectra
}
\details{
Many of
}
\examples{
data("raman_hdpe")

adj_intens(raman_hdpe)

}
\seealso{
\code{\link{subtr_bg}()} for spectral background correction;
\code{\link{match_spec}()} matches spectra with the Open Specy or other
reference libraries
}
\author{
Win Cowger, Zacharias Steinmetz
}
