% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{adjust_negative}
\alias{adjust_negative}
\alias{make_relative}
\title{Normalization of spectral data}
\usage{
adjust_negative(x, na.rm = FALSE)

make_relative(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one by
\code{as.vector(x, "numeric")}.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
\code{adjust_negative()} and \code{make_relative()} return numeric vectors
with the normalized data.
}
\description{
\code{adjust_negative()} converts numeric values \code{x} < 1 into values
>= 1, keeping absolute differences between values by shifting intensity
values with the value of the smallest number.
\code{make_relative()} converts values \code{x} into relative values between
0 and 1 using the standard normalization equation.
If \code{na.rm} is \code{TRUE}, missing values are removed before the
computation proceeds.
}
\details{
\code{adjust_negative()} is used in Open Specy to avoid errors that could
arise from log transforming spectra when using
\code{\link{adjust_intensity}()} and other functions.
\code{make_relative()} is used in Open Specy to retain the relative
height proportions between spectra while avoiding the large numbers that can
result from some spectral instruments.
}
\examples{
adjust_negative(c(-1000, -1, 0, 1, 10))
make_relative(c(-1000, -1, 0, 1, 10))

}
\seealso{
\code{\link[base]{min}()} for the calculation of minima;
\code{\link{adjust_intensity}()} for log transformation functions
}
\author{
Win Cowger, Zacharias Steinmetz
}
