% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_OpenSpecy.R
\name{head.OpenSpecy}
\alias{head.OpenSpecy}
\alias{print.OpenSpecy}
\alias{plot.OpenSpecy}
\alias{lines.OpenSpecy}
\alias{summary.OpenSpecy}
\title{Generic Open Specy Functions}
\usage{
\method{head}{OpenSpecy}(x, ...)

\method{print}{OpenSpecy}(x, ...)

\method{plot}{OpenSpecy}(x, ...)

\method{lines}{OpenSpecy}(x, ...)

\method{summary}{OpenSpecy}(object, ...)
}
\arguments{
\item{x}{an OpenSpecy object.}

\item{object}{an OpenSpecy object.}

\item{\ldots}{args.}
}
\value{
\code{print()} returns a textual representation of an OpenSpecy object.
\code{head()} returns the first few lines of an OpenSpecy object.
}
\description{
Functions to manipulate and visualize OpenSpecy objects.
}
\details{
\code{print()} prints the contents of an OpenSpecy object.
\code{head()} shows the first few lines of an OpenSpecy object.
}
\examples{
data("raman_hdpe")

# Printing the OpenSpecy object
print(raman_hdpe)

# Displaying the first few lines of the OpenSpecy object
head(raman_hdpe)

# Plotting the spectra
plot(raman_hdpe)

}
\seealso{
Other related functions.
}
\author{
Zacharias Steinmetz, Win Cowger
}
