% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_timestamp.R
\name{human_timestamp}
\alias{human_timestamp}
\title{Create human readable timestamps}
\usage{
human_timestamp()
}
\description{
This helper function creates human readable timestamps in the form of
\code{\%Y\%m\%d-\%H\%M\%OS} at the current time.
}
\details{
Human readable timestamps are appended to file names and fields when metadata
are shared to the Open Specy data portal at
\url{https://www.dropbox.com/sh/vibzoiq1i4o5usp/AAAe50RJhnvt6NoMDAJZ1CY8a?dl=0}
}
\examples{
human_timestamp()

}
\seealso{
\code{\link[base]{format.Date}} for date conversion functions
}
