% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_timestamp.R
\name{human_timestamp}
\alias{human_timestamp}
\title{Create human readable timestamps}
\usage{
human_timestamp()
}
\description{
This helper function creates human readable timestamps in the form of
\code{\%Y\%m\%d-\%H\%M\%OS}.
}
\examples{
human_timestamp()

}
\seealso{
\code{\link[base]{format.Date}} for date conversion functions
}
