% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_spec.R
\name{write_spec}
\alias{write_spec}
\alias{write_spec.default}
\alias{write_spec.OpenSpecy}
\alias{read_spec}
\title{Read and write spectral data}
\usage{
write_spec(x, ...)

\method{write_spec}{default}(x, ...)

\method{write_spec}{OpenSpecy}(x, file, method = NULL, digits = getOption("digits"), ...)

read_spec(file, share = NULL, method = NULL, ...)
}
\arguments{
\item{x}{a list object of class \code{\link{OpenSpecy}}.}

\item{file}{file to be read from or written to.}

\item{method}{submethod to be used for reading text files; defaults to
\code{\link[data.table]{fread}()} but \code{\link[utils]{read.csv}()} works
as well.}

\item{digits}{number of significant digits to use when formatting numeric
values; defaults to \code{\link[base]{getOption}("digits")}.}

\item{share}{defaults to \code{NULL}; needed to share spectra with the
Open Specy community; see \code{\link{share_spec}()} for details.}

\item{\ldots}{further arguments passed to the submethods.}
}
\value{
\code{read_spec()} returns a list object of class \code{\link{OpenSpecy}()}
containing spectral data.
}
\description{
Functions for reading and writing spectral data from Open Specy .yaml, .json,
or .rds.
}
\details{
reference to qs
}
\examples{
read_spec(read_extdata("raman_hdpe.yml"))
read_spec(read_extdata("raman_hdpe.json"))
read_spec(read_extdata("raman_hdpe.rds"))

\dontrun{
write_spec(raman_hdpe, "raman_hdpe.yml")
write_spec(raman_hdpe, "raman_hdpe.json")
write_spec(raman_hdpe, "raman_hdpe.rds")
}

}
\seealso{
\code{\link{OpenSpecy}()}
\code{\link[base]{saveRDS}()}; \code{\link[base]{readRDS}()};
\code{\link[yaml]{write_yaml}()}; \code{\link[yaml]{read_yaml}()};
\code{\link[jsonlite]{write_json}()}; \code{\link[jsonlite]{read_json}()};
}
\author{
Zacharias Steinmetz, Win Cowger
}
