% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_lib.R
\name{get_lib}
\alias{get_lib}
\alias{load_lib}
\title{Manage spectral libraries}
\usage{
get_lib(which = c("ftir", "raman"), conflicts = "overwrite", ...)

load_lib(which = c("ftir", "raman"), ...)
}
\arguments{
\item{which}{which}

\item{conflicts}{determines what happens when a file with the same name
exists at the specified destination. Can be one of the following (see
\code{\link[osfr]{osf_download}()} for details):
\itemize{
\item "error": throw an error and abort the file transfer operation.
\item "skip": skip the conflicting file(s) and continue transferring the remaining files.
\item "overwrite" (the default): replace the existing file with the transferred copy.
}}

\item{\ldots}{arguments passed to \code{\link[osfr]{osf_download}()}}
}
\description{
Description
}
\examples{
\dontrun{
get_lib()
}

}
\seealso{
seealso
}
