% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_spec.R
\name{c_spec}
\alias{c_spec}
\title{Manage spectral objects}
\usage{
c_spec(objects, wavenumber_transform = NULL, res = 5)
}
\arguments{
\item{objects}{a list object of class \code{OpenSpecy}.}

\item{wavenumber_transform}{a numeric if providing your own wavenumbers or character argument of "common_range" depending on how you want to merge the wavenumbers between the spectra, NULL will be interpreted as the spectra are all the same wavenumber range, see details for more information.}

\item{res}{defaults to \code{NULL}, the resolution you want the output wavenumbers to be.}
}
\value{
A single OpenSpecy object.
}
\description{
Functions for
}
\details{
details
}
\examples{
spectra <- lapply(c(read_extdata("raman_hdpe.csv"),  read_extdata("ftir_ldpe_soil.asp")), read_any)
spectra2 <- c_spec(objects = spectra, wavenumber_transform = "common_range", res = 5)
spectra3 <- c_spec(objects = spectra, wavenumber_transform = c(1000, 2000), res = 5)

}
\seealso{
\code{\link[OpenSpecy]{conform_spec}()};
\code{\link[OpenSpecy]{conform_res}()};
\code{\link[OpenSpecy]{as_OpenSpecy}()};
}
\author{
Zacharias Steinmetz, Win Cowger
}
