% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_spec.R
\name{conform_spec}
\alias{conform_spec}
\alias{c_spec}
\title{Manage spectral objects}
\usage{
conform_spec(spec, x, xout)

c_spec(...)
}
\arguments{
\item{x}{a list object of class \code{OpenSpecy}.}

\item{file}{file to be read from or written to.}

\item{share}{defaults to \code{NULL}; needed to share spectra with the
Open Specy community; see \code{\link{share_spec}()} for details.}

\item{method}{submethod to be used for reading text files; defaults to
\code{\link[data.table]{fread}()} but \code{\link[utils]{read.csv}()} works
as well.}

\item{\ldots}{further arguments passed to the submethods.}
}
\value{
All \code{read_*()} functions return data frames containing two columns
named \code{"wavenumber"} and \code{"intensity"}.
}
\description{
Functions for
}
\details{
details
}
\examples{
c()

}
\seealso{
\code{\link[hyperSpec]{read.jdx}()};
}
\author{
Zacharias Steinmetz, Win Cowger
}
