% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_spec.R
\name{cor_spec}
\alias{cor_spec}
\alias{cor_spec.default}
\alias{cor_spec.OpenSpecy}
\alias{ident_spec}
\alias{get_metadata}
\alias{get_metadata.default}
\alias{get_metadata.OpenSpecy}
\alias{max_cor_named}
\alias{filter_spec}
\alias{filter_spec.default}
\alias{filter_spec.OpenSpecy}
\title{Identify and filter spectra}
\usage{
cor_spec(x, ...)

\method{cor_spec}{default}(x, ...)

\method{cor_spec}{OpenSpecy}(x, library, na.rm = T, ...)

ident_spec(
  cor_matrix,
  x,
  library,
  top_n = NULL,
  add_library_metadata = NULL,
  add_object_metadata = NULL,
  ...
)

get_metadata(x, ...)

\method{get_metadata}{default}(x, ...)

\method{get_metadata}{OpenSpecy}(x, logic, rm_empty = TRUE, ...)

max_cor_named(cor_matrix, na.rm = T)

filter_spec(x, ...)

\method{filter_spec}{default}(x, ...)

\method{filter_spec}{OpenSpecy}(x, logic, ...)
}
\arguments{
\item{x}{an OpenSpecy object, typically with unknowns.}

\item{library}{An OpenSpecy object representing the library of spectra to correlate with.}

\item{na.rm}{Logical value indicating whether missing values should be removed when calculating correlations. Default is \code{TRUE}.}

\item{cor_matrix}{A correlation matrix for object and library, can be returned by \code{cor_spec()}}

\item{top_n}{Integer value specifying the number of top matches to return. If NULL (default), all matches will be returned.}

\item{add_library_metadata}{Name of the column in the library metadata containing the column names or NULL if you don't want to join.}

\item{add_object_metadata}{Name of the column in the object metadata containing the column names or NULL if you don't want to join.}

\item{logic}{a logical or numeric vector describing which spectra to keep (TRUE).}

\item{rm_empty}{Whether to remove empty columns in the metadata where there are no values.}

\item{\ldots}{Additional arguments passed to the \code{cor()} function for correlation calculation.}
}
\value{
A data table containing correlations between spectra and the library.
The table has three columns: \code{object_id}, \code{library_id}, and \code{match_val}.
Each row represents a unique pairwise correlation between a spectrum in the object and a spectrum in the library.
If \code{top_n} is specified, only the top \code{top_n} matches for each object spectrum will be returned.
If \code{add_library_metadata} is \code{is.character}, the library metadata will be added to the output.
If \code{add_object_metadata} is \code{is.character}, the object metadata will be added to the output.
}
\description{
This function correlates two OpenSpecy objects, typically one with knowns and one with unknowns.
}
\examples{
data("test_lib")
unknown <- read_any(read_extdata("ftir_ldpe_soil.asp")) |>
  conform_spec(new_wavenumbers = test_lib$wavenumber,
               res = spec_res(test_lib)) |>
  process_spec()
matches <- cor_spec(unknown, test_lib)

test_lib_extract <- filter_spec(test_lib,
  logic = grepl("polycarbonate", test_lib$metadata$polymer_class,
                ignore.case = TRUE)
)

matches2 <- cor_spec(unknown, library = test_lib_extract)

}
\seealso{
\code{\link{adj_intens}()} converts spectra;
\code{\link{get_lib}()} retrieves the Open Specy reference library;
\code{\link{load_lib}()} loads the Open Specy reference library into an \R
object of choice
}
\author{
Win Cowger, Zacharias Steinmetz
}
