% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_spectrum.R
\name{match_spectrum}
\alias{match_spectrum}
\title{Match spectra with reference library}
\usage{
match_spectrum(
  spectrum,
  library,
  which = NULL,
  type = "full",
  range = seq(0, 6000, 0.1),
  col_names = NULL,
  top_n = 100,
  ...
)
}
\arguments{
\item{spectrum}{spectrum}

\item{library}{library}

\item{which}{which}

\item{type}{type}

\item{range}{range}

\item{col_names}{col_names}

\item{top_n}{top_n}

\item{\ldots}{...}
}
\description{
desc
}
\examples{
data("raman_hdpe")


}
\seealso{
seealso
}
