% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenSpecy-package.R
\docType{package}
\name{OpenSpecy-package}
\alias{OpenSpecy}
\alias{OpenSpecy-package}
\title{OpenSpecy: Analyze, Process, Identify, and Share, Raman and (FT)IR Spectra}
\description{
Raman and (FT)IR spectral analysis tool for plastic particles and
    other environmental samples. Supported features include reading spectral
    data files (*.asp, *.csv, *.jdx, *.spc, *.spa, *.0), smoothing spectral
    intensities with `smooth_intensity()`, correcting background scatter with 
    `subtract_background()`, and identifying spectra using an onboard reference
    library. Analyzed spectra can be shared with the Open Specy community.
    A Shiny app is available via `run_app()` or online at
    <http://www.openspecy.org>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wincowgerDEV/OpenSpecy}
  \item Report bugs at \url{https://github.com/wincowgerDEV/OpenSpecy/issues}
}

}
\author{
\strong{Maintainer}: Win Cowger \email{wincowger@gmail.com} (\href{https://orcid.org/0000-0001-9226-3104}{ORCID})

Authors:
\itemize{
  \item Zacharias Steinmetz \email{steinmetz-z@uni-landau.de} (\href{https://orcid.org/0000-0001-6675-5033}{ORCID})
}

Other contributors:
\itemize{
  \item Andrew Gray (\href{https://orcid.org/0000-0003-2252-7367}{ORCID}) [contributor]
  \item Hannah Hapich (\href{https://orcid.org/0000-0003-0000-6632}{ORCID}) [contributor]
  \item Jennifer Lynch (\href{https://orcid.org/0000-0003-3572-8782}{ORCID}) [contributor, data contributor]
  \item Hannah De Frond (\href{https://orcid.org/0000-0003-1199-0727}{ORCID}) [contributor, data contributor]
  \item Keenan Munno (\href{https://orcid.org/0000-0003-2916-5944}{ORCID}) [contributor, data contributor]
  \item Chelsea Rochman (\href{https://orcid.org/0000-0002-7624-711X}{ORCID}) [contributor, data contributor]
  \item Sebastian Primpke (\href{https://orcid.org/0000-0001-7633-8524}{ORCID}) [contributor, data contributor]
  \item Orestis Herodotou [contributor, data contributor]
}

}
\keyword{internal}
