% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_reports.R
\name{plot_OpenSpecy}
\alias{plot_OpenSpecy}
\alias{interactive_plot}
\title{Plots for Open Specy objects}
\usage{
plot_OpenSpecy(x, x2 = NULL, ...)

interactive_plot(x, selected_spectrum, x2 = NULL, selected_spectrum2 = NULL)
}
\arguments{
\item{x}{An OpenSpecy object containing metadata and spectral data for the first group.}

\item{x2}{An optional second OpenSpecy object containing metadata and spectral data for the second group.}

\item{selected_spectrum}{Optional index of the selected spectrum to highlight on the heatmap.}

\item{object}{An OpenSpecy object containing metadata and spectral data.}

\item{z}{Optional numeric vector specifying the intensity values for the heatmap. If not provided, the function will use the intensity values from the OpenSpecy object.}

\item{sn}{Optional numeric value specifying the signal-to-noise ratio threshold. If provided along with \code{min_sn}, regions with SNR below the threshold will be excluded from the heatmap.}

\item{cor}{Optional numeric value specifying the correlation threshold. If provided along with \code{min_cor}, regions with correlation below the threshold will be excluded from the heatmap.}

\item{min_sn}{Optional numeric value specifying the minimum signal-to-noise ratio for inclusion in the heatmap. Regions with SNR below this threshold will be excluded.}

\item{min_cor}{Optional numeric value specifying the minimum correlation for inclusion in the heatmap. Regions with correlation below this threshold will be excluded.}
}
\value{
A plotly heatmap object displaying the OpenSpecy data. A subplot containing the heatmap and spectra plot. A plotly object displaying the spectra from the OpenSpecy object(s).
}
\description{
These functions generate heatmaps, spectral plots, and interactive plots for OpenSpecy data.
}
\examples{
data("raman_hdpe")
tiny_map <- read_zip(read_extdata("CA_tiny_map.zip"))
plot_OpenSpecy(raman_hdpe)

correlation <- correlate_spectra(
  conform_spec(raman_hdpe, new_wavenumbers = raman_hdpe$wavenumber, res = 5),
  conform_spec(tiny_map, tiny_map$wavenumbers, res = 5)
)
heatmap_OpenSpecy(tiny_map, z = tiny_map$metadata$y)

sample_spec(tiny_map, size = 12) |>
  interactive_plot(selected_spectrum = 2, x2 = raman_hdpe)

}
\author{
Win Cowger, Zacharias Steinmetz
}
