% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_reports.R
\name{plot_OpenSpecy}
\alias{plot_OpenSpecy}
\alias{interactive_plot}
\title{Plots for Open Specy objects}
\usage{
plot_OpenSpecy(x, x2 = NULL, ...)

interactive_plot(x, selected_spectrum, x2 = NULL, selected_spectrum2 = NULL)
}
\arguments{
\item{x}{An OpenSpecy object containing metadata and spectral data for the first group.}

\item{x2}{An optional second OpenSpecy object containing metadata and spectral data for the second group.}

\item{selected_spectrum}{Optional index of the selected spectrum to highlight on the heatmap.}

\item{object}{An OpenSpecy object containing metadata and spectral data.}

\item{z}{Optional numeric vector specifying the intensity values for the heatmap. If not provided, the function will use the intensity values from the OpenSpecy object.}

\item{sn}{Optional numeric value specifying the signal-to-noise ratio threshold. If provided along with \code{min_sn}, regions with SNR below the threshold will be excluded from the heatmap.}

\item{cor}{Optional numeric value specifying the correlation threshold. If provided along with \code{min_cor}, regions with correlation below the threshold will be excluded from the heatmap.}

\item{min_sn}{Optional numeric value specifying the minimum signal-to-noise ratio for inclusion in the heatmap. Regions with SNR below this threshold will be excluded.}

\item{min_cor}{Optional numeric value specifying the minimum correlation for inclusion in the heatmap. Regions with correlation below this threshold will be excluded.}
}
\value{
A plotly heatmap object displaying the OpenSpecy data. A subplot containing the heatmap and spectra plot. A plotly object displaying the spectra from the OpenSpecy object(s).
}
\description{
These functions generate heatmaps, spectral plots, and interactive plots for OpenSpecy data.
}
\examples{
data1 <- read_any(read_extdata("raman_hdpe.json"))
data2 <- read_zip("inst/extdata/CA_tiny_map.zip")
plot_OpenSpecy(data1)
plot_OpenSpecy(data1, x2 = data2)
correlation <- correlate_spectra(conform_spec(data1, new_wavenumbers = data2$wavenumber, res = 5), conform_spec(data2, data2$wavenumbers, res = 5))
heatmap_OpenSpecy(data2, z = data2$metadata$y)
interactive_plot(x = data2, selected_spectrum = 5)
interactive_plot(data2, selected_spectrum = 2, x2 = data1, selected_spectrum2 = 1)

}
\author{
Win Cowger, Zacharias Steinmetz
}
