% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spectra.R
\name{process_spectra}
\alias{process_spectra}
\title{Adjust spectral intensities to absorbance units}
\usage{
process_spectra(
  object,
  active_preprocessing = T,
  range_decision = F,
  min_range = 0,
  max_range = 6000,
  carbon_dioxide_decision = F,
  carbon_dioxide_min = 2200,
  carbon_dioxide_max = 2420,
  smooth_decision = F,
  smooth_polynomial = 3,
  smooth_window = 11,
  baseline_decision = F,
  baseline_selection = "Polynomial",
  raw_baseline = F,
  baseline_polynomial = 8,
  wavenumber_fit = NULL,
  intensity_fit = NULL,
  derivative_decision = T,
  derivative_order = 1,
  derivative_polynomial = 3,
  abs = T,
  derivative_window = 11
)
}
\arguments{
\item{object}{a list object of class \code{OpenSpecy}.}

\item{type}{a character string specifying whether the input spectrum is
in absorbance units (\code{"none"}, default) or needs additional conversion
from \code{"reflectance"} or \code{"transmittance"} data.}

\item{make_rel}{logical; if \code{TRUE} spectra are automatically normalized
with \code{\link{make_rel}()}.}

\item{\ldots}{further arguments passed to submethods; this is
to \code{\link{adj_neg}()} for \code{adj_intens()} and
to \code{\link{conform_res}()} for \code{conform_intens()}.}
}
\value{
\code{adj_intens()} returns a data frame containing two columns
named \code{"wavenumber"} and \code{"intensity"}.
}
\description{
Converts reflectance or transmittance intensity units to absorbance units.
}
\details{
Many of the Open Specy functions will assume that the spectrum is in
absorbance units. For example, see \code{\link{match_spec}()} and
\code{\link{subtr_bg}()}.
To run those functions properly, you will need to first convert any spectra
from transmittance or reflectance to absorbance using this function.
The transmittance adjustment uses the \eqn{log10(1 / T)}
calculation which does not correct for system and particle characteristics.
The reflectance adjustment uses the Kubelka-Munk equation
\eqn{(1 - R)^2 / 2R}. We assume that the reflectance intensity
is a percent from 1-100 and first correct the intensity by dividing by 100
so that it fits the form expected by the equation.
}
\examples{
data("raman_hdpe")

adj_intens(raman_hdpe)

test_noise = as_OpenSpecy(x = seq(400,4000, by = 10), spectra = data.table(intensity = rnorm(361)))
test_noise = process_spectra(test_noise, range_decision = T, min_range = 1000, max_range = 3000, carbon_dioxide_decision = T, abs = F)
ggplot() +
geom_line(aes(x = test_noise$wavenumber, y = test_noise$spectra[[1]]))
}
\seealso{
\code{\link{subtr_bg}()} for spectral background correction;
\code{\link{match_spec}()} matches spectra with the Open Specy or other
reference libraries
}
\author{
Win Cowger, Zacharias Steinmetz
}
