% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman_hdpe.R
\docType{data}
\name{raman_hdpe}
\alias{raman_hdpe}
\title{Sample Raman spectrum}
\format{
A data table containing 964 rows and 2 columns:
\tabular{ll}{
\code{wavenumber}: \tab spectral wavenumber [1/cm] \cr
\code{intensity}: \tab absorbance values [-] \cr
}
}
\description{
Raman spectrum of high-density polyethylene (HDPE).
}
\references{
Cowger W, Gray A, Christiansen SH, Christiansen SH, Christiansen SH,
De Frond H, Deshpande AD, Hemabessiere L, Lee E, Mill L, et al. (2020).
“Critical Review of Processing and Classification Techniques for Images and
Spectra in Microplastic Research.” \emph{Applied Spectroscopy},
\strong{74}(9), 989–1010.
}
\author{
Win Cowger
}
\keyword{data}
