% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_envi.R
\name{read_envi}
\alias{read_envi}
\title{Read ENVI data}
\usage{
read_envi(
  file,
  header = NULL,
  share = NULL,
  metadata = list(file_name = basename(file), user_name = NULL, contact_info = NULL,
    organization = NULL, citation = NULL, spectrum_type = NULL, spectrum_identity = NULL,
    material_form = NULL, material_phase = NULL, material_producer = NULL,
    material_purity = NULL, material_quality = NULL, material_color = NULL,
    material_other = NULL, cas_number = NULL, instrument_used = NULL,
    instrument_accessories = NULL, instrument_mode = NULL, spectral_resolution = NULL,
    laser_light_used = NULL, number_of_accumulations = NULL, 
    
    total_acquisition_time_s = NULL, data_processing_procedure = NULL,
    level_of_confidence_in_identification = NULL, other_info = NULL, license =
    "CC BY-NC"),
  ...
)
}
\arguments{
\item{file}{name of the binary file.}

\item{header}{name of the ASCII header file. If \code{NULL}, the name of the
header file is guessed by looking for a second file with the same basename as
\code{file} but with .hdr extension.}

\item{share}{defaults to \code{NULL}; needed to share spectra with the
Open Specy community; see \code{\link{share_spec}()} for details.}

\item{metadata}{a named list of the metadata; see \code{\link{read_text}()}
for details.}

\item{\ldots}{further arguments passed to the submethods.}
}
\value{
An \code{OpenSpecy} object.
}
\description{
This function allows ENVI data import.
}
\details{
ENVI data usually consists of two files, an ASCII header and a binary data
file. The header contains all information necessary for correctly reading
the binary file via \code{\link[caTools]{read.ENVI}()}.
}
\seealso{
\code{\link[caTools]{read.ENVI}()}
}
\author{
Zacharias Steinmetz, Claudia Beleites
}
