% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multi.R
\name{read_any}
\alias{read_any}
\alias{read_zip}
\title{Read spectral data from multiple files}
\usage{
read_any(file, ...)

read_zip(file, ...)
}
\arguments{
\item{file}{file to be read from or written to.}

\item{\ldots}{further arguments passed to the submethods.}
}
\value{
All \code{read_*()} functions return data frames containing two columns
named \code{"wavenumber"} and \code{"intensity"}.
}
\description{
Wrapper functions for reading
}
\details{
x
}
\examples{
read_text(read_extdata("raman_hdpe.csv"))
read_asp(read_extdata("ftir_ldpe_soil.asp"))
read_opus(read_extdata("ftir_ps.0"))

}
\author{
Zacharias Steinmetz, Win Cowger
}
