% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus}
\alias{read_opus}
\title{Read spectral data from Bruker OPUS binary files}
\usage{
read_opus(
  file,
  share = NULL,
  metadata = list(file_name = basename(file), user_name = NULL, contact_info = NULL,
    organization = NULL, citation = NULL, spectrum_type = NULL, spectrum_identity = NULL,
    material_form = NULL, material_phase = NULL, material_producer = NULL,
    material_purity = NULL, material_quality = NULL, material_color = NULL,
    material_other = NULL, cas_number = NULL, instrument_used = NULL,
    instrument_accessories = NULL, instrument_mode = NULL, spectral_resolution = NULL,
    laser_light_used = NULL, number_of_accumulations = NULL, 
    
    total_acquisition_time_s = NULL, data_processing_procedure = NULL,
    level_of_confidence_in_identification = NULL, other_info = NULL, license =
    "CC BY-NC"),
  type = "spec",
  simplify = FALSE,
  wns_digits = 1L,
  atm_comp_minus4offset = FALSE
)
}
\arguments{
\item{file}{Character vector with path to file(s).}

\item{share}{defaults to \code{NULL}; needed to share spectra with the
Open Specy community; see \code{\link{share_spec}()} for details.}

\item{metadata}{a named list of the metadata; see details below.}

\item{type}{Character vector of spectra types to extract from OPUS binary
file. Default is \code{"spec"}, which will extract the final spectra, e.g.
expressed in absorbance (named \code{AB} in Bruker OPUS programs). Possible
additional values for the character vector supplied to \code{type} are
\code{"spec_no_atm_comp"} (spectrum of the sample without compensation for
atmospheric gases, water vapor and/or carbon dioxide),
\code{"sc_sample"} (single channel spectrum of the sample measurement), \code{"sc_ref"}
(single channel spectrum of the reference measurement),
\code{"ig_sample"} (interferogram of the sample measurement) and \code{"ig_ref"}
(interferogram of the reference measurement).}

\item{simplify}{Logical (defaults to \code{FALSE}): if set to \code{TRUE}, returns a
flattened list.
The first element of that list (\code{wavenumbers}) is the wavenumbers of the
first file read.
The second element (\code{spectra}) is a matrix of the corresponding spectra.
Especially useful when passing more than one file to the \code{file} option, for
example to read a suite of spectral file directly into a matrix.
If the files do not have the same number of wavebands, the wavebands of the
very first file passed to \code{file} is used as a reference for linear interpolation.}

\item{wns_digits}{Integer that specifies the number of decimal places used to
round the wavenumbers (values of x-variables) if \code{simplify = TRUE}.}

\item{atm_comp_minus4offset}{Logical whether spectra after atmospheric
compensation are read with an offset of -4 bytes from Bruker OPUS files.
Default is \code{FALSE}.}
}
\value{
The output of \code{read_opus()} depends on the value of the \code{simplify}
option used in the function call.

If \code{simplify = FALSE} (default), \code{read_opus()} returns a list of 10 elements:

\describe{
\item{\code{metadata}}{a \code{data.frame} containing metadata from the OPUS file.}
\item{\code{spec}}{if \code{"spec"} was requested in the \code{type} option, a matrix of
the spectrum of the sample (otherwise set to \code{NULL}).}
\item{\code{spec_no_atm_comp}}{if \code{"spec_no_atm_comp"} was requested in the
\code{type} option, a matrix of the spectrum of the sample without background
compensation (otherwise set to \code{NULL}).}
\item{\code{sc_sample}}{if \code{"sc_sample"} was requested in the \code{type} option,
a matrix of the single channel spectrum of the sample (otherwise set to
\code{NULL}).}
\item{\code{sc_ref}}{if \code{"sc_ref"} was requested in the \code{type} option, a matrix
of the single channel spectrum of the reference (otherwise set to \code{NULL}).}
\item{\code{ig_sample}}{if \code{"ig_sample"} was requested in the \code{type} option, a
matrix of the interferogram of the sample (otherwise set to \code{NULL}).}
\item{\code{ig_ref}}{if \code{"ig_ref"} was requested in the \code{type} option, a matrix
of the interferogram of the reference (otherwise set to \code{NULL}).}
\item{\code{wavenumbers}}{if \code{"spec"} or \code{"spec_no_atm_comp"} was requested in
the \code{type} option, a numeric vector of the wavenumbers of the spectrum of
the sample (otherwise set to \code{NULL}).}
\item{\code{wavenumbers_sc_sample}}{if \code{"sc_sample"} was requested in the \code{type}
option, a numeric vector of the wavenumbers of the single channel
spectrum of the sample (otherwise set to \code{NULL}).}
\item{\code{wavenumbers_sc_ref}}{if \code{"sc_ref"} was requested in the \code{type}
option, a numeric vector of the wavenumbers of the single channel spectrum
of the reference (otherwise set to \code{NULL}).}
}

If \code{simplify = TRUE}, a list of two elements is returned:

\describe{
\item{\code{wavenumbers}}{numeric vector with wavenumbers of the requested
spectra.}
\item{\code{spectra}}{matrix with spectra of requested type (see argument
\code{type}).}
}
}
\description{
Read file(s) acquired with a Bruker Vertex FTIR Instrument.
}
\details{
The type of spectra returned by the function when using
\code{type = "spec"} depends on the setting of the Bruker instrument: typically,
it can be either absorbance or reflectance.

The type of spectra to extract from the file can also use Bruker's OPUS
software naming conventions, as follows:
\itemize{
\item \code{ScSm} corresponds to \code{sc_sample}
\item \code{ScRf} corresponds to \code{sc_ref}
\item \code{IgSm} corresponds to \code{ig_sample}
\item \code{IgRf} corresponds to \code{ig_ref}
}
}
\examples{
data <- read_opus(read_extdata("ftir_ps.0"))

}
\author{
Philipp Baumann, Zacharias Steinmetz, Win Cowger
}
