% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Open Specy app}
\usage{
run_app(...)
}
\arguments{
\item{\dots}{arguments passed to \code{\link[shiny]{runApp}}.}
}
\description{
This function starts the graphical user interface of Open Specy.
}
\examples{
\dontrun{
run_app(launch.browser = TRUE)
}

}
