% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Open Specy app}
\usage{
run_app(app_dir = system.file("shiny", package = "OpenSpecy"), ...)
}
\arguments{
\item{app_dir}{the app to run; defaults to
\code{system.file("shiny", package = "OpenSpecy")}.}

\item{\dots}{arguments passed to \code{\link[shiny]{runApp}()}.}
}
\value{
This function normally does not return any value, see
\code{\link[shiny]{runApp}()}.
}
\description{
This wrapper function starts the graphical user interface of Open Specy.
}
\details{
After running this function the Open Specy GUI should open in a separate
window or in your computer browser.
}
\examples{
\dontrun{
run_app()
}

}
\seealso{
\code{\link[shiny]{runApp}()}
}
\author{
Zacharias Steinmetz
}
