% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_intensity.R
\name{smooth_intensity}
\alias{smooth_intensity}
\alias{smooth_intensity.formula}
\alias{smooth_intensity.data.frame}
\alias{smooth_intensity.default}
\title{Adjust spectral intensities}
\usage{
smooth_intensity(x, ...)

\method{smooth_intensity}{formula}(formula, data = NULL, ...)

\method{smooth_intensity}{data.frame}(x, ...)

\method{smooth_intensity}{default}(x, y, p = 3, n = 11, make_relative = TRUE, ...)
}
\arguments{
\item{x}{wavenumber}

\item{formula}{formula}

\item{data}{data}

\item{y}{intensity}

\item{p}{p}

\item{n}{n}

\item{make_relative}{make_relative}

\item{\ldots}{...}
}
\description{
This smoother can enhance the signal to noise ratio of the data and uses a Savitzky-Golay filter with 12 running data points and the polynomial specified.
}
\examples{
data("raman_hdpe")
smooth_intensity(intensity ~ wavenumber, data = raman_hdpe)

}
\seealso{
seealso
}
