% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_res.R
\name{spec_res}
\alias{spec_res}
\title{Spectral resolution}
\usage{
spec_res(x)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one by
\code{as.vector(x, "numeric")}; \code{x} should be \code{wavenumber} data.}
}
\value{
\code{spec_res()} returns a single numeric value.
}
\description{
Helper function for calculating the spectral resolution from
\code{wavenumber} data.
}
\details{
The spectral resolution is the the minimum wavenumber, wavelength, or
frequency difference between two lines in a spectrum that can still be
distinguished.
}
\examples{
data("raman_hdpe")
spec_res(raman_hdpe$wavenumber)

}
\author{
Win Cowger, Zacharias Steinmetz
}
