% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_background.R
\name{subtract_background}
\alias{subtract_background}
\alias{subtract_background.formula}
\alias{subtract_background.default}
\title{Automated background subtraction for spectral data}
\usage{
subtract_background(x, ...)

\method{subtract_background}{formula}(formula, data = NULL, ...)

\method{subtract_background}{default}(x, y, degree = 8, ...)
}
\arguments{
\item{x}{wavenumber}

\item{formula}{formula}

\item{data}{data}

\item{y}{absorbance}

\item{degree}{the degree of the polynomial. Must be less than the number of
unique points when raw is false, as by default.}

\item{\ldots}{...}
}
\description{
This baseline correction routine iteratively finds the baseline of a spectrum
using a polynomial fitting.
}
\examples{
data("raman_hdpe")
subtract_background(absorbance ~ wavenumber, data = raman_hdpe)

}
\seealso{
seealso
}
