% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_lib.R
\docType{data}
\name{test_lib}
\alias{test_lib}
\title{Test reference library}
\format{
A list named \code{"test"} with two elements:

\tabular{ll}{
\code{metadata}: \tab metadata of 34 Raman spectra \cr
\code{library}: \tab all reference spectra, \code{sample_name} serves as
identifier \cr
}
}
\description{
Reference library of 34 Raman spectra used for internal testing.
}
\author{
Jennifer Lynch
}
\keyword{data}
