% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_lib.R
\docType{data}
\name{test_lib}
\alias{test_lib}
\title{Test reference library}
\format{
An \code{OpenSpecy} object; \code{sample_name} is the class of the spectra.
}
\description{
Reference library with 29 FTIR and 28 Raman spectra used for examples and
internal testing.
}
\examples{
data("test_lib")

}
\author{
Win Cowger
}
\keyword{data}
