% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_lib.R
\docType{data}
\name{test_lib}
\alias{test_lib}
\title{Test reference library}
\format{
A lsit
\tabular{ll}{
wavenumber: \tab \cr
absorbance: \tab \cr
}
}
\description{
Desc
}
\references{
ref
}
\author{
author
}
\keyword{data}
